Create QuickSort
The Quicksort Algorithm
This Algorithm selects an element as pivot element from the given array and partitions the array around it such that, Left side of pivot contains all the elements that are less than the pivot element. The right side contains all elements that are greater than the pivot element. Let P be the index of pivot after partitioning the array. Then, the left subarray(start to P-1) and right subarray(P+1 to end) are sorted recursively to get the final sorted array as output.
Partitioning is the process of placing the pivot in the just right ordered place.

How the array is partitioned around pivot?
Assume the last element as pivot .
We want larger elements to go to the right and smaller to go to left. We maintain two indexes i and j, whereas i pointing to the leftmost element and j is pointing to last element(pivot).
While i is to the left of j, we move i right, skipping all the elements less than the pivot. If an element is found greater than the pivot, i stops
While j is to the right of i, we move j left, skipping all the elements greater than the pivot. If an element is found less than the pivot, j stops
When both i and j have stopped, the elements are swapped.
When i and j have crossed, no swap is performed, scanning stops, and the element pointed to by i is swapped with the pivot.